+++ inst/examples/00-test-script.R
@@ -2,5 +2,8 @@ library(leaflet)
 
 leaflet() %>%
   addProviderTiles("CartoDB.Positron") %>%
-  addTargomoPolygons()
+  addTargomoPolygons(lat = 51.5, lng = -0.18,
+                     options = targomoOptions(travelType = "transit"),
+                     color = c("red", "orange", "green")) %>%
+  addMarkers(lat = 51.5, lng = -0.18)